requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
cloudinary==1.44.1
dj-database-url==3.0.0
Django==5.2.3
django-cloudinary-storage==0.3.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
